Move pagination hook and styles to a shared util
