+++ public/app/plugins/panel/state-timeline/StateTimelinePanel.tsx
@@ -1,12 +1,9 @@
-import { css } from '@emotion/css';
 import { useMemo, useState } from 'react';
-import { useMeasure } from 'react-use';
 
-import { DashboardCursorSync, DataFrame, PanelProps } from '@grafana/data';
+import { DashboardCursorSync, PanelProps } from '@grafana/data';
 import {
   AxisPlacement,
   EventBusPlugin,
-  Pagination,
   TooltipDisplayMode,
   TooltipPlugin2,
   usePanelContext,
@@ -15,7 +12,6 @@ import {
 import { TimeRange2, TooltipHoverMode } from '@grafana/ui/src/components/uPlot/plugins/TooltipPlugin2';
 import { TimelineChart } from 'app/core/components/TimelineChart/TimelineChart';
 import {
-  makeFramePerSeries,
   prepareTimelineFields,
   prepareTimelineLegendItems,
   TimelineMode,
@@ -26,72 +22,10 @@ import { OutsideRangePlugin } from '../timeseries/plugins/OutsideRangePlugin';
 import { getTimezones } from '../timeseries/utils';
 
 import { StateTimelineTooltip2 } from './StateTimelineTooltip2';
-import { Options, defaultOptions } from './panelcfg.gen';
+import { Options } from './panelcfg.gen';
+import { containerStyles, usePagination } from './utils';
 
 interface TimelinePanelProps extends PanelProps<Options> {}
-
-const styles = {
-  container: css({
-    display: 'flex',
-    flexDirection: 'column',
-  }),
-  paginationContainer: css({
-    display: 'flex',
-    justifyContent: 'center',
-    width: '100%',
-  }),
-  paginationElement: css({
-    marginTop: '8px',
-  }),
-};
-
-function usePagination(frames?: DataFrame[], perPage?: number) {
-  const [currentPage, setCurrentPage] = useState(1);
-
-  const [paginationWrapperRef, { height: paginationHeight, width: paginationWidth }] = useMeasure<HTMLDivElement>();
-
-  const pagedFrames = useMemo(
-    () => (!perPage || frames == null ? frames : makeFramePerSeries(frames)),
-    [frames, perPage]
-  );
-
-  if (!perPage || pagedFrames == null) {
-    return {
-      paginatedFrames: pagedFrames,
-      paginationRev: 'disabled',
-      paginationElement: undefined,
-      paginationHeight: 0,
-    };
-  }
-
-  perPage ||= defaultOptions.perPage!;
-
-  const numberOfPages = Math.ceil(pagedFrames.length / perPage);
-  // `perPage` changing might lead to temporarily too large values of `currentPage`.
-  const currentPageCapped = Math.min(currentPage, numberOfPages);
-  const pageOffset = (currentPageCapped - 1) * perPage;
-  const currentPageFrames = pagedFrames.slice(pageOffset, pageOffset + perPage);
-
-  // `paginationRev` needs to change value whenever any of the pagination settings changes.
-  // It's used in to trigger a reconfiguration of the underlying graphs (which is cached,
-  // hence an explicit nudge is required).
-  const paginationRev = `${currentPageCapped}/${perPage}`;
-
-  const showSmallVersion = paginationWidth < 550;
-  const paginationElement = (
-    <div className={styles.paginationContainer} ref={paginationWrapperRef}>
-      <Pagination
-        className={styles.paginationElement}
-        currentPage={currentPageCapped}
-        numberOfPages={numberOfPages}
-        showSmallVersion={showSmallVersion}
-        onNavigate={setCurrentPage}
-      />
-    </div>
-  );
-
-  return { paginatedFrames: currentPageFrames, paginationRev, paginationElement, paginationHeight };
-}
 
 /**
  * @alpha
@@ -141,7 +75,7 @@ export const StateTimelinePanel = ({
   const enableAnnotationCreation = Boolean(canAddAnnotations && canAddAnnotations());
 
   return (
-    <div className={styles.container}>
+    <div className={containerStyles.container}>
       <TimelineChart
         theme={theme}
         frames={paginatedFrames}
